indexes: refactoring and fixes applying changes to 0.13
